requirements.txt
certifi==2019.3.9
chardet==3.0.4
Django==2.1.7
django-filter==2.1.0
django-rest-framework==0.1.0
djangorestframework==3.9.2
gunicorn==19.9.0
idna==2.8
Pillow==5.4.1
psycopg2==2.7.7
pytz==2018.9
requests==2.21.0
urllib3==1.24.1
psycopg2-binary
